Add error handling (#6)
In the initial release, I forgot to hook up the error streams from
nyxx_extensions and nyxx_commands to our bot. This meant some
interactions with the bot would just fail silently if an error occurred.

This PR adds error handling to respond to users even if an error occurs
with an error message.